Module type BatEnum.Enumerable

A signature for data structures which may be converted to and from enum.

If you create a new data structure, you should make it compatible with Enumerable.

type 'a enumerable

The data structure, e.g. 'a List.t

val enum : 'a enumerable -> 'a t

Return an enumeration of the elements of the data structure

val of_enum : 'a t -> 'a enumerable

Build a data structure from an enumeration