Module BatHashcons.H

Hashing utilities

val hc0_ : int -> int

hc0_ h corresponds to the hashcode of a first constructor applied to an object of hashcode h

val hc0 : 'a hobj -> int

hc0 ho is the hashcode of a first constructor applied to the hashed object ho

val hc1_ : int -> int -> int

hc1_ h k corresponds to the hashcode of the kth constructor applied to an object of hashcode h.

val hc1 : 'a hobj -> int -> int

hc1 ho k corresponds to the hashcode of the kth constructor applied to the hashed object ho.