Module S.Infix

Infix operators over a BatHashtbl

val (-->) : 'a t -> key -> 'a

tbl-->x returns the current binding of x in tbl, or raises Not_found if no such binding exists. Equivalent to Hashtbl.find tbl x

val (<--) : 'a t -> (key * 'a) -> unit

tbl<--(x, y) adds a binding of x to y in table tbl. Previous bindings for x are not removed, but simply hidden. That is, after performing Hashtbl.remove tbl x, the previous binding for x, if any, is restored. (Same behavior as with association lists.) Equivalent to Hashtbl.add tbl x y