Module BatIMap.Infix

Infix operators over a BatIMap

val (-->) : 'a t -> int -> 'a

map-->key returns the current binding of key in map, or

raises Not_found
if no such binding exists. Equivalent to find key map.

val (<--) : 'a t -> (int * 'a) -> 'a t

map<--(key, value) returns a map containing the same bindings as map, plus a binding of key to value. If key was already bound in map, its previous binding disappears. Equivalent to add key value map

Important warning: BatIMap.add takes an optional argument, eq that is missing in this operator <--. As a consequence, using <-- implies the use of Pervasives.(==) as comparison function.