Module BatLazyList.Labels

Operations on LazyList with labels.

This module overrides a number of functions of List by functions in which some arguments require labels. These labels are there to improve readability and safety and to let you change the order of arguments to functions. In every case, the behavior of the function is identical to that of the corresponding function of LazyList.

val iter : f:('a -> 'b) -> 'a t -> unit

val iteri : f:(int -> 'a -> unit) -> 'a t -> unit

val map : f:('a -> 'b) -> 'a t -> 'b t

val mapi : f:(int -> 'a -> 'b) -> 'a t -> 'b t

val fold_left : f:('a -> 'b -> 'a) -> init:'a -> 'b t -> 'a

val fold_right : f:('a -> 'b -> 'b) -> init:'b -> 'a t -> 'b

val find : f:('a -> bool) -> 'a t -> 'a

val rfind : f:('a -> bool) -> 'a t -> 'a

val find_exn : f:('a -> bool) -> exn -> 'a t -> 'a

val rfind_exn : f:('a -> bool) -> exn -> 'a t -> 'a

val findi : f:(int -> 'a -> bool) -> 'a t -> int * 'a

val rfindi : f:(int -> 'a -> bool) -> 'a t -> int * 'a

val remove_if : f:('a -> bool) -> 'a t -> 'a t

val remove_all_such : f:('a -> bool) -> 'a t -> 'a t

val take_while : f:('a -> bool) -> 'a t -> 'a t

val drop_while : f:('a -> bool) -> 'a t -> 'a t

val filter : f:('a -> bool) -> 'a t -> 'a t

val exists : f:('a -> bool) -> 'a t -> bool

val for_all : f:('a -> bool) -> 'a t -> bool

val filter_map : f:('a -> 'b option) -> 'a t -> 'b t

val map2 : f:('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val iter2 : f:('a -> 'b -> unit) -> 'a t -> 'b t -> unit

val fold_right2 : f:('a -> 'b -> 'c -> 'c) -> 'a t -> 'b t -> init:'c -> 'c

val for_all2 : f:('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val exists2 : f:('a -> 'b -> bool) -> 'a t -> 'b t -> bool

module Exceptionless : sig ... end