Module BatList.Labels

Operations on List with labels.

This module overrides a number of functions of List by functions in which some arguments require labels. These labels are there to improve readability and safety and to let you change the order of arguments to functions. In every case, the behavior of the function is identical to that of the corresponding function of List.

val init : int -> f:(int -> 'a) -> 'a list

val iter : f:('a -> unit) -> 'a list -> unit

val iteri : f:(int -> 'a -> unit) -> 'a list -> unit

val map : f:('a -> 'b) -> 'a list -> 'b list

val mapi : f:(int -> 'a -> 'b) -> 'a list -> 'b list

val rev_map : f:('a -> 'b) -> 'a list -> 'b list

val fold_left : f:('a -> 'b -> 'a) -> init:'a -> 'b list -> 'a

val fold : f:('a -> 'b -> 'a) -> init:'a -> 'b list -> 'a

val fold_right : f:('a -> 'b -> 'b) -> 'a list -> init:'b -> 'b

val iter2 : f:('a -> 'b -> unit) -> 'a list -> 'b list -> unit

val map2 : f:('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val rev_map2 : f:('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val fold_left2 : f:('a -> 'b -> 'c -> 'a) -> init:'a -> 'b list -> 'c list -> 'a

val fold_right2 : f:('a -> 'b -> 'c -> 'c) -> 'a list -> 'b list -> init:'c -> 'c

val for_all : f:('a -> bool) -> 'a list -> bool

val exists : f:('a -> bool) -> 'a list -> bool

val for_all2 : f:('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val exists2 : f:('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val subset : cmp:('a -> 'b -> int) -> 'a list -> 'b list -> bool

val find : f:('a -> bool) -> 'a list -> 'a

val find_exn : f:('a -> bool) -> exn -> 'a list -> 'a

val find_map_opt : f:('a -> 'b option) -> 'a list -> 'b option

val findi : f:(int -> 'a -> bool) -> 'a list -> int * 'a

val rfind : f:('a -> bool) -> 'a list -> 'a

val filter : f:('a -> bool) -> 'a list -> 'a list

val filter_map : f:('a -> 'b option) -> 'a list -> 'b list

val count_matching : f:('a -> bool) -> 'a list -> int

val concat_map : f:('a -> 'b list) -> 'a list -> 'b list

val find_all : f:('a -> bool) -> 'a list -> 'a list

val partition : f:('a -> bool) -> 'a list -> 'a list * 'a list

val partition_map : f:('a -> ('b, 'c) BatEither.t) -> 'a list -> 'b list * 'c list

val remove_if : f:('a -> bool) -> 'a list -> 'a list

val take_while : f:('a -> bool) -> 'a list -> 'a list

val drop_while : f:('a -> bool) -> 'a list -> 'a list

val stable_sort : ?cmp:('a -> 'a -> int) -> 'a list -> 'a list

val fast_sort : ?cmp:('a -> 'a -> int) -> 'a list -> 'a list

val merge : ?cmp:('a -> 'a -> int) -> 'a list -> 'a list -> 'a list

module LExceptionless : sig ... end