Module BatLog.Easy

A simple-to-use logger with verbosity levels that outputs by default to stderr (changeable at runtime) with the date and time at the beginning of each log message.

val level : easy_lev ref

Set this ref to the lowest level of log you want logged. For example, Easy.level := `always disables all logging except that at the `always level. Setting Easy.level := `info will enable logging for `info, `warn, `error, `fatal and `always levels.

val output : unit BatIO.output ref

Set this ref to the output you want logging messages to go to. Defaults to stderr.

val log : ?fp:string -> easy_lev -> string -> unit

log lev msg logs the message msg if the current logging level is lev or lower.

val logf : ?fp:string -> easy_lev -> ('a, unit BatIO.output, unit) format -> 'a

As log, but instead of a string message, a printf format is allowed with whatever arguments are appropriate.