Module BatLog.Make

Build a logger module with custom, fixed output, prefix and flags

Parameters

module S : Config

Signature

val log : ?fp:string -> string -> unit

print s logs the message s, returning unit.

val logf : ?fp:string -> ('a, S.t BatIO.output, unit) Pervasives.format -> 'a

As Printf.printf, only the message is printed to the logging output and prefixed with status information per the current flags and the currently set prefix.

val fatal : ?fp:string -> string -> 'a

fatal s logs the message s and then calls exit 1. This exits the program with return code 1.

val fatalf : ?fp:string -> ('a, S.t BatIO.output, unit) Pervasives.format -> 'a

fatalf allows a format string (as Printf.printf)and the arguments to that format string to build the logging message. Exits the program with return code 1.