Module type BatLog.Level_sig

The details of a level scheme for verbosity-level loggers

type t

A type for level values, usually a polymorphic variant

val to_string : t -> string

Convert each level to a string

val default_level : t

The default level for loggers created with this; log messages with level less than this won't be printed by default.

val compare : t -> t -> int

a comparison function between levels, to know whether logging at a particular level should be printed