Module Float.Infix

Infix operators over a BatMap

val (-->) : 'a t -> key -> 'a

map-->key returns the current binding of key in map, or raises Not_found. Equivalent to find key map.

val (<--) : 'a t -> (key * 'a) -> 'a t

map <-- (key, value) returns a map containing the same bindings as map, plus a binding of key to value. If key was already bound in map, its previous binding disappears. Equivalent to add key value map.