Module type BatNumber.Compare

And if you are ready to drop generic comparison operators, then you can open this one as well

type bat__compare_t

val (<>) : bat__compare_t -> bat__compare_t -> bool

val (>=) : bat__compare_t -> bat__compare_t -> bool

val (<=) : bat__compare_t -> bat__compare_t -> bool

val (>) : bat__compare_t -> bat__compare_t -> bool

val (<) : bat__compare_t -> bat__compare_t -> bool

val (=) : bat__compare_t -> bat__compare_t -> bool