Module BatOptParse

Modules for GNU getopt(3)-style command line parsing.

author Bardur Arantsson

module Opt : sig ... end

This module contains the basic functions and types for defining new option types and accessing the values of options.

module StdOpt : sig ... end

This module contains various standard options.

module Formatter : sig ... end

This module contains the types and functions for implementing custom usage message formatters.

module OptParser : sig ... end

This module contains the option parser itself.