Module OfString.OperatorLift

Convenience operator for lifting primitive strings to ustring type.

@documents Future.Path.OperatorLift

val (!!) : string -> ustring

Prefix operator that converts primitive string to ustring. May raise some exceptions depending on actual strings implementation.

You might want to open Path.OperatorLift to improve readability of path construction using string literals. Example: Path.root/:!!"foo"/:!!"bar" = Path.root/:(S.lift "foo")/:(S.lift "bar") (where S.lift converts to ustring type)