Module Incubator.Private_state_enums

module State : sig ... end

val enum_bits : unit -> int BatEnum.t

These enumerations are built on a copy of the global RNG state. To keep successive constructions from using the same RNG state, when any of these functions is called, the global RNG state is perturbed by using its current internal state as seed to construct a new state.

val enum_int : int -> int BatEnum.t

val enum_bool : unit -> bool BatEnum.t

val enum_float : float -> float BatEnum.t

val enum_int32 : Int32.t -> Int32.t BatEnum.t

val enum_int64 : Int64.t -> Int64.t BatEnum.t

val enum_nativeint : Nativeint.t -> Nativeint.t BatEnum.t

val enum_char : unit -> char BatEnum.t