Module Int.Labels

Operations on Set with labels.

This module overrides a number of functions of Set by functions in which some arguments require labels. These labels are there to improve readability and safety and to let you change the order of arguments to functions. In every case, the behavior of the function is identical to that of the corresponding function of Set.

val iter : f:(elt -> unit) -> t -> unit

val fold : f:(elt -> 'a -> 'a) -> t -> init:'a -> 'a

val for_all : f:(elt -> bool) -> t -> bool

val exists : f:(elt -> bool) -> t -> bool

val map : f:(elt -> elt) -> t -> t

val filter : f:(elt -> bool) -> t -> t

val filter_map : f:(elt -> elt option) -> t -> t

val partition : f:(elt -> bool) -> t -> t * t