Module BatSet.PSet

Polymorphic sets

The definitions below describe the polymorphic set interface.

They are similar in functionality to the functorized BatSet.Make module, but the compiler cannot ensure that sets using different element ordering have different types: the responsibility of not mixing non-sensical comparison functions together is to the programmer. If you ever need a custom comparison function, it is recommended to use the BatSet.Make functor for additional safety.

author Nicolas Cannasse

author Markus Mottl

author David Rajchenbach-Teller

type 'a t

The type of sets.

include BatEnum.Enumerable with type 'a enumerable = 'a t

type 'a enumerable = 'a t

The data structure, e.g. 'a List.t

include BatInterfaces.Mappable with type 'a mappable = 'a t

type 'a mappable = 'a t

The data structure, e.g. 'a List.t

val empty : 'a t

The empty set, using compare as comparison function

val create : ('a -> 'a -> int) -> 'a t

Creates a new empty set, using the provided function for key comparison.

val is_empty : 'a t -> bool

Test whether a set is empty or not.

val singleton : ?cmp:('a -> 'a -> int) -> 'a -> 'a t

Creates a new set with the single given element in it.

val mem : 'a -> 'a t -> bool

mem x s tests whether x belongs to the set s.

val find : 'a -> 'a t -> 'a

find x s returns the element in s that tests equal to x under its comparison function.

raises Not_found
if no element is equal

val find_opt : 'a -> 'a t -> 'a option

find_opt x s returns Some k for the element k in s that tests equal to x under its comparison function. If no element is equal, return None

val find_first : ('a -> bool) -> 'a t -> 'a

find_first f m returns the first element e for which f e is true or raises Not_found if there is no such element. f must be monotonically increasing, i.e. if k1 < k2 && f k1 is true then f k2 must also be true.

val find_first_opt : ('a -> bool) -> 'a t -> 'a option

find_first_opt f m returns Some e for the first element e for which f e is true or returns None if there is no such element. f must be monotonically increasing, i.e. if k1 < k2 && f k1 is true then f k2 must also be true.

val find_last : ('a -> bool) -> 'a t -> 'a

find_last f m returns the last element e for which f e is true or raises Not_found if there is no such element. f must be monotonically decreasing, i.e. if k1 < k2 && f k2 is true then f k1 must also be true.

val find_last_opt : ('a -> bool) -> 'a t -> 'a option

find_last_opt f m returns Some e for the last element e for which f e is true or returns None if there is no such element. f must be monotonically decreasing, i.e. if k1 < k2 && f k2 is true then f k1 must also be true.

val add : 'a -> 'a t -> 'a t

add x s returns a set containing all elements of s, plus x. If x was already in s, s is returned unchanged.

before 3.3.0
Physical equality was not ensured.

val remove : 'a -> 'a t -> 'a t

remove x s returns a set containing all elements of s, except x. If x was not in s, s is returned unchanged.

before 3.3.0
Physical equality was not ensured.

val remove_exn : 'a -> 'a t -> 'a t

remove_exn x s behaves like remove x s except that it raises an exception if x is not in s.

raises Not_found
if x is not in s.

val update : 'a -> 'a -> 'a t -> 'a t

update x y s replace x by y in s. update is faster when x compares equal to y according to the comparison function used by your set. When x and y are physically equal, m is returned unchanged.

raises Not_found
if x is not in s.

before 3.3.0
Physical equality was not ensured.

val union : 'a t -> 'a t -> 'a t

union s t returns the union of s and t - the set containing all elements in either s and t. The returned set uses t's comparison function. The current implementation works better for small s.

val intersect : 'a t -> 'a t -> 'a t

intersect s t returns a new set of those elements that are in both s and t. The returned set uses s's comparison function.

val diff : 'a t -> 'a t -> 'a t

diff s t returns the set of all elements in s but not in t. The returned set uses s's comparison function.

val sym_diff : 'a t -> 'a t -> 'a t

sym_diff s t returns the set of all elements in s or t but not both. This is the same as diff (union s t) (inter s t). The returned set uses s's comparison function.

val compare : 'a t -> 'a t -> int

Total ordering between sets. Can be used as the ordering function for doing sets of sets.

val equal : 'a t -> 'a t -> bool

equal s1 s2 tests whether the sets s1 and s2 are equal, that is, contain equal elements.

val subset : 'a t -> 'a t -> bool

subset a b returns true if a is a subset of b. O(|a|).

val disjoint : 'a t -> 'a t -> bool

disjoint s1 s2 tests whether the sets s1 and s2 contain no shared elements. (i.e. inter s1 s2 is empty.)

val iter : ('a -> unit) -> 'a t -> unit

iter f s applies f in turn to all elements of s. The elements of s are presented to f in increasing order with respect to the ordering over the type of the elements.

val at_rank_exn : int -> 'a t -> 'a

at_rank_exn i s returns element at rank i in s, that is the i-th element in increasing order (the 0-th element being the smallest element of s).

raises Not_found
if s = empty.

raises Invalid_argument
error_message if i < 0 || i >= cardinal s

val map : ('a -> 'b) -> 'a t -> 'b t

map f x creates a new set with elements f a0, f a1... f aN, where a0, a1, ..., aN are the values contained in x

The resulting map uses the polymorphic compare function to order elements.

val map_endo : ('a -> 'a) -> 'a t -> 'a t

map f x creates a new set with elements f a0, f a1... f aN, where a0, a1, ..., aN are the values contained in x

The resulting map uses the same compare function to order elements as m does.

If f returns physically equal values for all elements in m then the resulting map will be physically equal to m.

val filter : ('a -> bool) -> 'a t -> 'a t

filter p s returns the set of all elements in s that satisfy predicate p.

if p returns true for all elements then s is returned unmodified.

before 3.3.0
Physical equality was not ensured.

val filter_map : ('a -> 'b option) -> 'a t -> 'b t

filter_map f m combines the features of filter and map. It calls calls f a0, f a1, f aN where a0,a1..an are the elements of m and returns the set of pairs bi such as f ai = Some bi (when f returns None, the corresponding element of m is discarded).

The resulting map uses the polymorphic compare function to order elements.

val filter_map_endo : ('a -> 'a option) -> 'a t -> 'a t

filter_map_endo f m combines the features of filter and map. It calls calls f a0, f a1, f aN where a0,a1..an are the elements of m and returns the set of pairs bi such as f ai = Some bi (when f returns None, the corresponding element of m is discarded).

The resulting map uses the same compare function to order elements as used for m.

if the filter function f returns true for all elements in m, the resulting map is physically equal to m.

val fold : ('a -> 'b -> 'b) -> 'a t -> 'b -> 'b

fold f s a computes (f xN ... (f x1 (f x0 a))...), where x0,x1..xN are the elements of s, in increasing order.

val exists : ('a -> bool) -> 'a t -> bool

exists p s checks if at least one element of the set satisfies the predicate p.

val for_all : ('a -> bool) -> 'a t -> bool

Returns whether the given predicate applies to all elements in the set

val partition : ('a -> bool) -> 'a t -> 'a t * 'a t

returns two disjoint subsets, those that satisfy the given predicate and those that don't

val split : 'a -> 'a t -> 'a t * bool * 'a t

split x s returns a triple (l, present, r), where l is the set of elements of s that are strictly less than x; r is the set of elements of s that are strictly greater than x; present is false if s contains no element equal to x, or true if s contains an element equal to x.

val split_opt : 'a -> 'a t -> 'a t * 'a option * 'a t

split_opt x s returns a triple (l, maybe_v, r), where l is the set of elements of s that are strictly less than x; r is the set of elements of s that are strictly greater than x; maybe_v is None if s contains no element equal to x, or Some v if s contains an element v that compares equal to x.

val split_lt : 'a -> 'a t -> 'a t * 'a t

split_lt x s returns a pair of sets (l, r), such that l is the subset of s with elements < x; r is the subset of s with elements >= x.

val split_le : 'a -> 'a t -> 'a t * 'a t

split_le x s returns a pair of sets (l, r), such that l is the subset of s with elements <= x; r is the subset of s with elements > x.

val cardinal : 'a t -> int

Return the number of elements of a set.

val elements : 'a t -> 'a list

Return the list of all elements of the given set. The returned list is sorted in increasing order with respect to the ordering of the given set.

val to_list : 'a t -> 'a list

Alias for elements.

val to_array : 'a t -> 'a array

Same as to_list but with an array instead of a list.

val min_elt : 'a t -> 'a

returns the smallest element of the set.

raises Not_found
if given an empty set.

val min_elt_opt : 'a t -> 'a option

Return Some e for the smallest element e of the given set (with respect to the Ord.compare ordering). Return None if the set is empty.

val pop_min : 'a t -> 'a * 'a t

Returns the smallest element of the given set along with the rest of the set. Semantically equivalent and faster than

let mini = min_elt s in (mini, remove mini s)

raises Not_found
if the set is empty.

val pop_max : 'a t -> 'a * 'a t

Returns the biggest element of the given set along with the rest of the set. Semantically equivalent and faster than

let maxi = max_elt s in (maxi, remove maxi s)

raises Not_found
if the set is empty.

val max_elt : 'a t -> 'a

returns the largest element of the set.

raises Not_found
if given an empty set.

val max_elt_opt : 'a t -> 'a option

Same as Set.S.min_elt_opt, but for the largest element of the given set.

val choose : 'a t -> 'a

returns an arbitrary (but deterministic) element of the given set.

raises Not_found
if given an empty set.

val choose_opt : 'a t -> 'a option

Return Some e for one element e of the given set. Which element is chosen is unspecified, but equal elements will be chosen for equal sets. Return None if the set is empty.

val any : 'a t -> 'a

Return one element of the given set. The difference with choose is that there is no guarantee that equals elements will be picked for equal sets. This merely returns the quickest element to get (O(1)).

raises Not_found
if the set is empty.

val pop : 'a t -> 'a * 'a t

returns one element of the set and the set without that element.

raises Not_found
if given an empty set

val enum : 'a t -> 'a BatEnum.t

Return an enumeration of all elements of the given set. The returned enumeration is sorted in increasing order with respect to the ordering of this set.

val of_enum : ?cmp:('a -> 'a -> int) -> 'a BatEnum.t -> 'a t

val of_enum_cmp : cmp:('a -> 'a -> int) -> 'a BatEnum.t -> 'a t

val of_list : ?cmp:('a -> 'a -> int) -> 'a list -> 'a t

builds a set from the given list, using the default comparison function

val of_array : ?cmp:('a -> 'a -> int) -> 'a array -> 'a t

builds a set from the given array and comparison function

val to_seq : 'a t -> 'a BatSeq.t

Iterate on the whole set, in ascending order.

val to_rev_seq : 'a t -> 'a BatSeq.t

Iterate on the whole set, in descending order.

val to_seq_from : 'a -> 'a t -> 'a BatSeq.t

to_seq_from x s iterates on a subset of the elements in s, namely those greater or equal to x, in ascending order.

val add_seq : 'a BatSeq.t -> 'a t -> 'a t

add the given elements to the set, in order.

val of_seq : ?cmp:('a -> 'a -> int) -> 'a BatSeq.t -> 'a t

build a set from the given elements

Boilerplate code

Printing

val print : ?first:string -> ?last:string -> ?sep:string -> ('a BatInnerIO.output -> 'c -> unit) -> 'a BatInnerIO.output -> 'c t -> unit

val get_cmp : 'a t -> 'a -> 'a -> int

get the comparison function used for a polymorphic map