Module Array.Exceptionless

Operations on Array without exceptions.

val find : ('a -> bool) -> 'a t -> 'a option

find p a returns Some x, where x is the first element of array a that satisfies the predicate p, or None if there is no such element.

val findi : ('a -> bool) -> 'a t -> int option

findi p a returns Some n, where n is the index of the first element of array a that satisfies the predicate p, or None if there is no such element.