Module Array.Labels

include module type of struct include BatArray.Labels end

Operations on Array with labels.

This module overrides a number of functions of Array by functions in which some arguments require labels. These labels are there to improve readability and safety and to let you change the order of arguments to functions. In every case, the behavior of the function is identical to that of the corresponding function of Array.

val init : int -> f:(int -> 'a) -> 'a array

val create : int -> init:'a -> 'a array

val make_matrix : dimx:int -> dimy:int -> 'a -> 'a array array

val create_matrix : dimx:int -> dimy:int -> 'a -> 'a array array

val sub : 'a array -> pos:int -> len:int -> 'a array

val fill : 'a array -> pos:int -> len:int -> 'a -> unit

val blit : src:'a array -> src_pos:int -> dst:'a array -> dst_pos:int -> len:int -> unit

val iter : f:('a -> unit) -> 'a array -> unit

val iteri : f:(int -> 'a -> unit) -> 'a array -> unit

val modify : f:('a -> 'a) -> 'a array -> unit

val modifyi : f:(int -> 'a -> 'a) -> 'a array -> unit

val fold_left : f:('a -> 'b -> 'a) -> init:'a -> 'b array -> 'a

val fold : f:('a -> 'b -> 'a) -> init:'a -> 'b array -> 'a

val fold_right : f:('b -> 'a -> 'a) -> 'b array -> init:'a -> 'a

val fold_while : p:('acc -> 'a -> bool) -> f:('acc -> 'a -> 'acc) -> init:'acc -> 'a array -> 'acc * int

val sort : cmp:('a -> 'a -> int) -> 'a array -> unit

val stable_sort : cmp:('a -> 'a -> int) -> 'a array -> unit

val fast_sort : cmp:('a -> 'a -> int) -> 'a array -> unit

val iter2 : f:('a -> 'b -> unit) -> 'a BatArray.t -> 'b BatArray.t -> unit

val exists : f:('a -> bool) -> 'a BatArray.t -> bool

val for_all : f:('a -> bool) -> 'a BatArray.t -> bool

val iter2i : f:(int -> 'a -> 'b -> unit) -> 'a BatArray.t -> 'b BatArray.t -> unit

val map : f:('a -> 'b) -> 'a BatArray.t -> 'b BatArray.t

val mapi : f:(int -> 'a -> 'b) -> 'a BatArray.t -> 'b BatArray.t

val filter : f:('a -> bool) -> 'a BatArray.t -> 'a BatArray.t

val filter_map : f:('a -> 'b option) -> 'a BatArray.t -> 'b BatArray.t

val count_matching : f:('a -> bool) -> 'a BatArray.t -> int

module LExceptionless = BatArray.Labels.LExceptionless

include module type of struct include BatArray.Labels.LExceptionless end

val find : f:('a -> bool) -> 'a BatArray.t -> 'a option

val findi : f:('a -> bool) -> 'a BatArray.t -> int option