Module Int32.Labels

Operations on Map with labels.

This module overrides a number of functions of Map by functions in which some arguments require labels. These labels are there to improve readability and safety and to let you change the order of arguments to functions. In every case, the behavior of the function is identical to that of the corresponding function of Map.

val add : key:key -> data:'a -> 'a t -> 'a t

val iter : f:(key:key -> data:'a -> unit) -> 'a t -> unit

val map : f:('a -> 'b) -> 'a t -> 'b t

val mapi : f:(key:key -> data:'a -> 'b) -> 'a t -> 'b t

val filterv : f:('a -> bool) -> 'a t -> 'a t

val filter : f:(key -> 'a -> bool) -> 'a t -> 'a t

val fold : f:(key:key -> data:'a -> 'b -> 'b) -> 'a t -> init:'b -> 'b

val compare : cmp:('a -> 'a -> int) -> 'a t -> 'a t -> int

val equal : cmp:('a -> 'a -> bool) -> 'a t -> 'a t -> bool