Module BatteriesHelp

Tools for reading the documentation from the toplevel

All these tools are invoked automatically by the Batteries Toplevel. They are provided here if you wish to integrate them into your own toplevel.

author David Teller

type kinds =

| Values

| Types

| Topics

| Modules

| Exns

| Modtypes

| Classes

| Methods

| Attributes

| Objtypes

val init : unit -> unit

Proceed to initialization.

This function loads the primary help files and registers the toplevel directives.

If you integrate the on-line help system into your toplevel, you must call this function before any of the other functions of this module.

val help : unit -> unit

help () opens the tutorial.

val man : string -> unit

man "something" opens the help about subject "something".